Move core dask method to base simulation.
